auto playlist use negative longs as ID, instead of hashcode

+++ features/detail/src/main/java/dev/olog/feature/detail/detail/DetailFragmentPresenter.kt
@@ -16,6 +16,7 @@ internal class DetailFragmentPresenter @Inject constructor(
 
 ) {
 
+    // TODO refactor, use proper type system
     suspend fun removeFromPlaylist(
         item: DetailFragmentModel.PlaylistTrack
     ) {
@@ -31,7 +32,7 @@ internal class DetailFragmentPresenter @Inject constructor(
         } else {
             removeFromPlaylistUseCase(
                 RemoveFromPlaylistUseCase.Input(
-                playlistId, item.idInPlaylist.toLong(), playlistType
+                playlistId, item.idInPlaylist, playlistType
             ))
         }
     }
